server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Erick Yan <erick.yanzhang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
flask-migrate = "^4.0.7"
kafka-python = "^2.0.2"
six = "^1.16.0"


[tool.poetry.group.dev.dependencies]
alembic = "1.13.0"
blinker = "1.7.0"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "41.0.7"
flask-script = "2.0.6"
flask-sqlalchemy = "3.1.1"
greenlet = "3.0.1"
idna = "3.6"
importlib-metadata = "7.0.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
mako = "1.3.0"
markupsafe = "2.1.3"
psycopg2-binary = "2.9.9"
pycparser = "2.21"
pyjwt = "2.8.0"
python-dotenv = "1.0.0"
requests = "2.31.0"
sqlalchemy = "2.0.23"
typing-extensions = "4.8.0"
urllib3 = "2.1.0"
werkzeug = "3.0.1"
zipp = "3.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
